requirements.txt
python-telegram-bot>=20.7,<21.0
Pillow>=10.0.0,<12.0.0
PyPDF2>=3.0.0,<4.0.0
pdf2docx>=0.5.8,<0.6.0
python-dotenv>=1.0.0,<2.0.0
Pillow-HEIF>=0.7.0
pytesseract>=0.3.10
opencv-python>=4.8.0
imageio-ffmpeg>=0.4.8
PyMuPDF>=1.23.0,<2.0.0
openpyxl>=3.1.0,<4.0.0
python-pptx>=0.6.23,<1.0.0
